Replaced One Test Case from WebDriver.IO to Playwright: consent-tools-vanilla-opt-in.test.ts and Removed all WebDriver.IO Test Cases
