+++ packages/consent/consent-tools-integration-tests/src/tests/example.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test'
 
-test('Homepage has welcome message', async ({ page }) => {
+test('Example Test Case to check playwright is installed', async ({ page }) => {
   await page.goto('https://www.example.com/')
   await expect(page.locator('text=Example Domain')).toBeVisible()
 })
